Add bodybuilder methods for findByUri
